premier_geek/api/pyproject.toml
[tool.poetry]
name = "premier-geek-api"
version = "0.0.0"
description = "An API for the Premier Geek app backend"
authors = [
    "Sebastian Kucharzyk <kucharzyk.sebastian@gmail.com>",
]
readme = "README.md"
packages = [{include = "runtime"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
mangum = "^0.17.0"
cognitojwt = {extras = ["sync"], version = "^1.4.1"}
pydantic-settings = "^2.4.0"
requests = "^2.32.3"
boto3 = "^1.35.5"
boto3-stubs = {extras = ["ssm"], version = "^1.35.5"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
